util: move trim functions into util.
