+++ src/util.rs
@@ -317,6 +317,37 @@ pub fn chomp(mut data: &[u8]) -> &[u8] {
     data
 }
 
+/// Trim the leading whitespace
+pub fn trim_start(input: &[u8]) -> &[u8] {
+    let mut result = input;
+    while let Some(x) = result.first() {
+        if is_space(*x) {
+            result = &result[1..]
+        } else {
+            break;
+        }
+    }
+    result
+}
+
+/// Trim the trailing whitespace
+pub fn trim_end(input: &[u8]) -> &[u8] {
+    let mut result = input;
+    while let Some(x) = result.last() {
+        if is_space(*x) {
+            result = &result[..(result.len() - 1)]
+        } else {
+            break;
+        }
+    }
+    result
+}
+
+/// Trim the leading and trailing whitespace from this byteslice.
+pub fn trimmed(input: &[u8]) -> &[u8] {
+    trim_end(trim_start(input))
+}
+
 /// Determines if character is a whitespace character.
 /// whitespace = ' ' | '\t' | '\r' | '\n' | '\x0b' | '\x0c'
 pub fn is_space(c: u8) -> bool {
@@ -1461,6 +1492,17 @@ mod tests {
     #[case("te\nst", "te\nst")]
     fn test_chomp(#[case] input: &str, #[case] expected: &str) {
         assert_eq!(chomp(input.as_bytes()), expected.as_bytes());
+    }
+
+    #[rstest]
+    #[case::trimmed(b"notrim", b"notrim")]
+    #[case::trim_start(b"\t trim", b"trim")]
+    #[case::trim_both(b" trim ", b"trim")]
+    #[case::trim_both_ignore_middle(b" trim trim ", b"trim trim")]
+    #[case::trim_end(b"trim \t", b"trim")]
+    #[case::trim_empty(b"", b"")]
+    fn test_trim(#[case] input: &[u8], #[case] expected: &[u8]) {
+        assert_eq!(trimmed(input), expected);
     }
 
     #[rstest]
